prepared for training

+++ models/official/resnet/onecycle.py
@@ -46,9 +46,9 @@ class OneCycleScheduler():
 
     def getlrmom(self,step):
         if step<self.phase_1_steps:
-            return tf.constant(self.phases[0][0].getval(step),dtype=tf.float64),tf.constant(self.phases[0][1].getval(step),dtype=tf.float64)
+            return tf.constant(self.phases[0][0].getval(step),dtype=tf.float32),tf.constant(self.phases[0][1].getval(step),dtype=tf.float32)
         else:
-            return tf.constant(self.phases[1][0].getval(step-self.phase_1_steps),dtype=tf.float64),tf.constant(self.phases[1][1].getval(step-self.phase_1_steps),dtype=tf.float64)
+            return tf.constant(self.phases[1][0].getval(step-self.phase_1_steps),dtype=tf.float32),tf.constant(self.phases[1][1].getval(step-self.phase_1_steps),dtype=tf.float32)
 
 
 def lrs(step,total_step):
